Ignoring `_pypy_speedup.c` when regenerating.
This is also a temporary hack. See failure:
https://circleci.com/gh/dhermes/bezier/1466

The near term fix will be to generate both `_speedup.c` and
`_pypy_speedup.c` in the same script.

The long term fix will hopefully be a removal of `_pypy_speedup.c`
once PyPy releases a new version that supports latest NumPy.

+++ scripts/remove_cython_files.py
@@ -23,6 +23,7 @@ import subprocess
 
 _SCRIPTS_DIR = os.path.abspath(os.path.dirname(__file__))
 ROOT_DIR = os.path.dirname(_SCRIPTS_DIR)
+IGNORED = (os.path.join("src", "bezier", "_pypy_speedup.c"),)
 
 
 def main():
@@ -35,6 +36,8 @@ def main():
         return
 
     for filename in all_files.split("\n"):
+        if filename in IGNORED:
+            continue
         os.remove(filename)
